refactor(db): /schemas -> /database
